[sith]  Implement `sith` main module

+++ Libs/sith/Gameplay/sithTime.c
@@ -55,7 +55,7 @@ void sithTime_Advance(void)
 
     if ( (sithMain_g_sith_mode.debugModeFlags & SITHDEBUG_SLOWMODE) != 0 )
     {
-        sithTime_g_frameTime = rintf((float)sithTime_g_frameTime * 0.2f); // frameTime / 5.0f
+        sithTime_g_frameTime = (uint32_t)rintf((float)sithTime_g_frameTime * 0.2f); // frameTime / 5.0f
     }
 
     sithTime_g_frameTimeFlex = (float)sithTime_g_frameTime * 0.001f;
